Copy Zero-K/.gitattributes
Keeping in sync/comparing differences with upstream is a bit of a
pain without this.


From the ZK commit message (0efc50f9910fe66bc483337eca):

The standard setting for line conversion is autocrlf - on linux, this
defaults to false, and on windows, the installer defaults to true -
though this can be changed, if for example this is installed on a
windows VM where the underlying filesystem is shared with linux.

Given .h, .glsl, .fs, and .json files were not listed and seemingly able
to be worked with by the windows developers without complaint,
presumably they already have this configured correctly (being the
default), and these settings were only serving as an annoyance for linux
users.

Linux users may want to double check that they're at least set to use
core.autocrlf input, and possibly re-checkout files, run dos2unix, or
otherwise make sure that they have native line endings in their working
tree going forward.